test(shared-server): separate AppInbox handler owner suites
